refactor(cli): cmd/client 由 cobra 迁移至 lynx-go/commands v0.3.0

+++ cmd/client/cmd/groups_test.go
@@ -124,7 +124,7 @@ func TestBuildListMembershipsReq(t *testing.T) {
 		name      string
 		groupID   string
 		queries   string
-		pageSize  int32
+		pageSize  int
 		pageToken string
 		wantErr   string
 	}{
@@ -148,7 +148,7 @@ func TestBuildListMembershipsReq(t *testing.T) {
 				require.NotContains(t, req, "queries")
 			}
 			if tt.pageSize > 0 {
-				require.Equal(t, int32(10), req["pageSize"])
+				require.Equal(t, 10, req["pageSize"])
 			} else {
 				require.NotContains(t, req, "pageSize")
 			}
